Abstract Deposit fetching logic from RampRegistration
