+++ client/src/contexts/Deposits/index.ts
@@ -1,0 +1,2 @@
+export { default as DepositsContext } from './DepositsContext'
+export { default as DepositsProvider } from './DepositsProvider'
