+++ client/src/contexts/RampRegistration/RampRegistrationProvider.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState, ReactNode } from 'react'
 import { useContractRead } from 'wagmi'
 
-import { Deposit } from './types'
 import { ZERO_ADDRESS } from '../../helpers/constants'
 import useAccount from '@hooks/useAccount'
 import useSmartContracts from '@hooks/useSmartContracts';
@@ -13,7 +12,7 @@ interface ProvidersProps {
   children: ReactNode;
 }
 
-const AccountProvider = ({ children }: ProvidersProps) => {
+const RampRegistrationProvider = ({ children }: ProvidersProps) => {
   /*
    * Contexts
    */
@@ -25,7 +24,6 @@ const AccountProvider = ({ children }: ProvidersProps) => {
    */
   const [registrationHash, setRegistrationHash] = useState<string>("");
   const [registeredVenmoId, setRegisteredVenmoId] = useState<string>("");
-  const [deposits, setDeposits] = useState<Deposit[]>([]);
 
   /*
     Helpers
@@ -50,19 +48,6 @@ const AccountProvider = ({ children }: ProvidersProps) => {
     args: [loggedInEthereumAddress],
   })
 
-  // function getAccountDeposits(address _account) external view returns (Deposit[] memory accountDeposits) {
-  const {
-    data: depositsRaw,
-    isLoading: isFetchDepositsLoading,
-    isError: isFetchDepositsError,
-    // refetch: refetchDeposits,
-  } = useContractRead({
-    address: rampAddress,
-    abi: rampAbi,
-    functionName: 'getAccountDeposits',
-    args: [loggedInEthereumAddress],
-  })
-
   // mapping(bytes32 => Intent) public intents;
   // const {
   //   data: intentRaw,
@@ -82,9 +67,6 @@ const AccountProvider = ({ children }: ProvidersProps) => {
 
   // mapping(address => AccountInfo) public accounts;
   // mapping(bytes32 => bytes32) public venmoIdIntent;
-
-  // function getDeposit(uint256 _depositId) external view returns (Deposit memory) {
-  // function getDepositFromIds(uint256[] memory _depositIds) external view returns (Deposit[] memory depositArray) {
 
   /*
    * Hooks
@@ -117,41 +99,12 @@ const AccountProvider = ({ children }: ProvidersProps) => {
   //   }
   // }, [isLoggedIn, refetchVenmoIdHash]);
 
-  useEffect(() => {
-    if (!isFetchDepositsLoading && !isFetchDepositsError && depositsRaw) {
-      const depositsArrayRaw = depositsRaw as any[];
-
-      const sanitizedDeposits: Deposit[] = [];
-      for (let i = depositsArrayRaw.length - 1; i >= 0; i--) {
-        const depositData = depositsArrayRaw[i];
-        
-        const deposit: Deposit = {
-          depositor: depositData.depositor.toString(),
-          remainingDepositAmount: depositData.remainingDeposits.toString(),
-          outstandingIntentAmount: depositData.outstandingIntentAmount.toString(),
-          conversionRate: depositData.conversionRate.toString(),
-          convenienceFee: depositData.convenienceFee.toString(),
-        };
-
-        sanitizedDeposits.push(deposit);
-      }
-
-      if (isLoggedIn && depositsRaw) {
-        setDeposits(sanitizedDeposits);
-        console.log(sanitizedDeposits);
-      } else {
-        setDeposits([]);
-      }
-    }
-  }, [isLoggedIn, depositsRaw, isFetchDepositsLoading, isFetchDepositsError]);
-
   return (
     <RampRegistrationContext.Provider
       value={{
         isRegistered,
         registrationHash,
         registeredVenmoId,
-        deposits,
       }}
     >
       {children}
@@ -159,4 +112,4 @@ const AccountProvider = ({ children }: ProvidersProps) => {
   );
 };
 
-export default AccountProvider
+export default RampRegistrationProvider
