+++ client/src/hooks/useDeposits.ts
@@ -1,0 +1,9 @@
+import { useContext } from 'react'
+
+import { DepositsContext } from '../contexts/Deposits'
+
+const useDeposits = () => {
+  return { ...useContext(DepositsContext) }
+}
+
+export default useDeposits
